Refactoring File Reader.
