+++ src/main/java/livefilereader/ProcessOutputInfo.java
@@ -1,5 +1,7 @@
 package livefilereader;
 
+import java.io.File;
+
 /**
  * This interface provides line to byte offset translation that should be used to access the output capturing file
  * efficiently.
@@ -9,5 +11,7 @@ public interface ProcessOutputInfo {
 	public int getLineByteOffset(int line);
 
 	public int getLineEnd(int line);
+
+	public File getFile();
 
 }
