+++ src/main/java/livefilereader/ReaderUI.java
@@ -41,9 +41,15 @@ public class ReaderUI extends JFrame {
 					    jre + "/bin/java.exe", "-jar", projectDir + "/ConsoleTest.jar"
 					};
 
-					String[] command = commandRandom;
+					String[] commandService = new String[] {
+					    jre + "/bin/java.exe",
+					    "-Dorg.gradle.appname=user-service -classpath .\\gradle\\wrapper\\gradle-wrapper.jar org.gradle.wrapper.GradleWrapperMain bootrun"
+					};
+
+					String[] command = commandService;
 
 					ProcessBuilder builder = new ProcessBuilder(command);
+					builder.directory(new File(System.getProperty("user.home") + "/git/user-service"));
 					builder.redirectErrorStream(true);
 
 					ReaderController controller = new ReaderController();
